LEDs, Auto, tuning robot
The button board now will control the color of the LEDs on the robot (when they work), and can change the color between yellow, purple, and when disabled it will be orange.

Started to add offsets for the other swerve modules.

Removed testing keybindings and added Sendable Chooser for selecting auto routines.

Because the grabber placer has the new cube hat (the goober), in the ScoreGamePieces command, the grabber placer now runs again with a positive voltage to move it up and prevent it from yanking the game pieces off of nodes after scoring them.

We need to find the lowest possible working amp limit for the conveyor motor so we don't snap it again. We should also do this for our other motors.