test for tomorrow for cone + cube high + balance barrier
In auto for the cone + cube high + balance barrier auto routine, we frequently experience the cube for the second high score not being handed off to the placer grabber causing us to score one less game piece in auto. To fix this issue, we want to try running the indexer conveyor and walls backwards to spit out the cube after the elevator has started extending. This will either cause the cube to be placed to the high node if it gets handed to the grabber placer successfully, or if it misses, the intake and indexer will spit it out into the hybrid node.